+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to
   [#2528](https://github.com/OpenFn/lightning/issues/2528)
 - Fix editor panel title gets pushed away when collapsed
   [#2545](https://github.com/OpenFn/lightning/issues/2545)
+- Mark unfinished steps having finished runs as `lost`
+  [#2416](https://github.com/OpenFn/lightning/issues/2416)
 
 ## [v2.9.8] - 2024-10-03